Add scripting to transfer proxy settings to sandbox
This commit adds scripts to detect proxy environment variables on the host
and pass them via files to the sandbox container. It also copies those files
to the correct locations inside of the container.

This provides /etc/environment for programs that honor proxy environment
variables and it provides /root/.m2/settings.xml for Maven (which does not
honor the environment variables set for proxies)

There may still be a remaining issue with "make containers" inside the sandbox
requiring some additional proxy configuration.

Change-Id: I27d289f9f6cc0948489fc2544e3be4c90096dd33
Partial-bug: 1766986