Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> then it's better to place the utility into CI itself - contrail-zuul-jobs 
> This repo is not used in CI and it would be not so simple to create job that will clone this repo just for utility. Also it's not so correct.

Eventually it will be integrated with CI, but that's a more complicated task.  For now, it's useful for the code to be in a place where it can be run on demand to report on Python 3 readiness.

I have documented there are better ways to obtain this information but it is more cumbersome.  The script is only one file, and requires no changes to any components.  I'm not sure what makes the script seem big other than documentation.  I would not let the perfect be the enemy of the good.

Patch-set: 2
CC: Gerrit User 790 <790@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Reviewer: Gerrit User 822 <822@c72cd27d-1743-4a55-84a4-cd13b8400c52>