Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > then it's better to place the utility into CI itself - contrail-zuul-jobs 
> > This repo is not used in CI and it would be not so simple to create job that will clone this repo just for utility. Also it's not so correct.
> 
> Eventually it will be integrated with CI, but that's a more complicated task.  For now, it's useful for the code to be in a place where it can be run on demand to report on Python 3 readiness.
> 
> I have documented there are better ways to obtain this information but it is more cumbersome.  The script is only one file, and requires no changes to any components.  I'm not sure what makes the script seem big other than documentation.  I would not let the perfect be the enemy of the good.

I suppose that this repository will be dropped - All repos should be moved to TF soon but TF already have own dev-env.
I agree that it can be run from this place by developer. But this script is very similar to UnitTests - then why UT are placed with components and not separately? Also we have separate repo contrail-test with functional tests for all components...
This repo has different purpose than testing something - it's intended just for build.
Therefore I still suggest to move this script to CI infra or to some another repo with infra tools.

Patch-set: 2