Update patch set 2
Patch Set 2:

But why the tool should be placed somewhere as soon as possible? Would it be better to choose appropriate place and commit it there?
With monthly releases it will be common situation to be in the middle.

Also I do not think that packages like bash-completion can help for python3 development.

Patch-set: 2