Update patch set 12
Patch Set 12:

Hi Abhay, is it possible to approve this?  Prabhu was to approve but we had issues in CI and then Prabhu moved on.  Bartosz has approved it.  This is to make it possible to build a container on Mac in a contrail-dev-env.  I fixed it a long time ago.

Patch-set: 12
Reviewer: Gerrit User 106 <106@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Reviewer: Gerrit User 822 <822@c72cd27d-1743-4a55-84a4-cd13b8400c52>